Update access to location Get Current Weather and 5 days Historical weather

+++ WeatherApp/WeatherApp.Common/Models/API/WeatherApiModel.cs
@@ -2,7 +2,7 @@
 {
     public class WeatherApiModel
     {
-        public int lat { get; set; }
-        public int lon { get; set; }
+        public double lat { get; set; }
+        public double lon { get; set; }
     }
 }
